Remove ability to log all script users IPs
Using a custom URL instead of a direct GitHub link allows the host to log every script download and the source IP address. It also introduces a security risk: the server hosting the custom URL (ipregion.vrnt.xyz) could be compromised, and the script replaced with a malicious one. This is not possible when using a direct GitHub link